move CallParams to inside JibriSelenium
