+++ src/test/kotlin/org/jitsi/jibri/api/http/HttpApiTest.kt
@@ -36,7 +36,6 @@ import io.kotlintest.specs.ShouldSpec
 import org.glassfish.jersey.jackson.JacksonFeature
 import org.glassfish.jersey.server.ResourceConfig
 import org.glassfish.jersey.test.JerseyTest
-import org.jitsi.jibri.CallParams
 import org.jitsi.jibri.CallUrlInfo
 import org.jitsi.jibri.JibriManager
 import org.jitsi.jibri.RecordingSinkType
@@ -45,6 +44,7 @@ import org.jitsi.jibri.StartServiceResult
 import org.jitsi.jibri.config.XmppCredentials
 import org.jitsi.jibri.health.EnvironmentContext
 import org.jitsi.jibri.health.JibriHealth
+import org.jitsi.jibri.selenium.CallParams
 import javax.ws.rs.client.Entity
 import javax.ws.rs.core.Application
 import javax.ws.rs.ext.ContextResolver
